+++ MumbleSharp/Constants.cs
@@ -3,8 +3,8 @@ namespace MumbleSharp
 {
     public static class Constants
     {
-        public const ushort DEFAULT_AUDIO_SAMPLE_RATE = 48000;
-        public const ushort DEFAULT_AUDIO_SAMPLE_BITS = 16;
-        public const ushort DEFAULT_AUDIO_SAMPLE_CHANNELS = 1;
+        public const int DEFAULT_AUDIO_SAMPLE_RATE = 48000;
+        public const byte DEFAULT_AUDIO_SAMPLE_BITS = 16;
+        public const byte DEFAULT_AUDIO_SAMPLE_CHANNELS = 1;
     }
 }
