+++ MumbleSharp/Model/User.cs
@@ -39,7 +39,7 @@ namespace MumbleSharp.Model
 
         private readonly CodecSet _codecs;
 
-        public User(IMumbleProtocol owner, uint id, ushort audioSampleRate = Constants.DEFAULT_AUDIO_SAMPLE_RATE, ushort audioSampleBits = Constants.DEFAULT_AUDIO_SAMPLE_BITS, ushort audioSampleChannels = Constants.DEFAULT_AUDIO_SAMPLE_CHANNELS)
+        public User(IMumbleProtocol owner, uint id, int audioSampleRate = Constants.DEFAULT_AUDIO_SAMPLE_RATE, byte audioSampleBits = Constants.DEFAULT_AUDIO_SAMPLE_BITS, byte audioSampleChannels = Constants.DEFAULT_AUDIO_SAMPLE_CHANNELS)
         {
             _codecs = new CodecSet(audioSampleRate, audioSampleBits, audioSampleChannels);
             _buffer = new AudioDecodingBuffer(audioSampleRate, audioSampleBits, audioSampleChannels);
